Detect memory usage via /proc and shuffle things to disk accordingly.
The problem is that the OS never reclaims memory, and when seeing that
we're *still* over the memory limit, we just keep putting stuff on disk
until everything is there.  This is obviously a total fail.  But I'm
committing the code because it's closer to where we want to be than it
was before.